跟后台联调后的一些修正，转测版本B008。

+++ Models/Response/GetDishSaleInfoResponse.cs
@@ -21,7 +21,7 @@ namespace Models.Response
     {
         public string dishName { get; set; }
 
-        public int dishCount { get; set; }
+        public decimal dishCount { get; set; }
 
         public decimal? totlePrice { get; set; }
     }
